OSD: port upstream changes from Brain
